made a file for the app
