+++ app.js
@@ -1,0 +1,53 @@
+const morgan = require('morgan')
+const express = require('express')
+const app = express()
+const cors = require('cors')
+const mongoose = require('mongoose')
+const config = require('./utils/config')
+const Blog = require('./model/blog')
+const logger = require('./utils/logger')
+
+const mongoUrl = config.MONGODB_URL
+
+logger.info('connecting to', config.MONGODB_URL)
+
+mongoose
+  .connect(mongoUrl, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+    useCreateIndex: true,
+  })
+  .then(() => {
+    console.log('connected to MongoDB')
+  })
+  .catch((error) => {
+    console.log('error connecting to MongoDB:', error.message)
+  })
+
+app.use(cors())
+app.use(express.json())
+
+morgan.token('body', (req, res) => {
+  return JSON.stringify(req.body)
+})
+
+app.use(
+  morgan(':method :url :status :res[content-length] - :response-time ms :body')
+)
+
+app.get('/api/blogs', (request, response) => {
+  Blog.find({}).then((blogs) => {
+    response.json(blogs)
+  })
+})
+
+app.post('/api/blogs', (request, response) => {
+  const blog = new Blog(request.body)
+
+  blog.save().then((result) => {
+    response.status(201).json(result)
+  })
+})
+
+module.exports = app
