+++ package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "dotenv": "^9.0.2",
     "express": "^4.17.1",
+    "http": "*",
     "mongoose": "^5.12.9",
     "mongoose-unique-validator": "^2.0.3",
     "morgan": "^1.10.0"
